requirements.txt
streamlit
langchain
langchain-community
langchain-core
mysql-connector-python
python-dotenv
langchain-huggingface